+++ src/components/charts-generic/use-chart-state.tsx
@@ -93,6 +93,7 @@ export type GroupedBarsState = {
   colors: ScaleOrdinal<string, string>;
   opacityScale: ScaleOrdinal<string, number>;
   xAxisLabel?: string;
+  labelsBySegment: Record<string, { prefix: string; lines: string[] }>;
 };
 
 type BarsState = {